Move race car examples into car/racing
