+++ examples/car/racing/_tracks/readDataFcn.py
